% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{buildModel}
\alias{buildModel}
\title{Builds hydroState model}
\usage{
buildModel(
  input.data = data.frame(year = c(), flow = c(), precip = c()),
  data.transform = "",
  stateModel = "",
  markovModel = ""
)
}
\arguments{
\item{input.data}{dataframe of annual runoff and precipitation observations. For running seasonal models with \code{seasonal.parameters}, monthly data is required.}

\item{data.transform}{a \code{Qhat.object} with transformed observations from \code{select.transform}. If blank, the default uses 'boxcox' to transform observations.}

\item{stateModel}{a \code{QhatModel.object} from \code{select.stateModel}. If blank, the default selects a 2-state 'QhatModel.homo.normal.linear' model with a truncated normal error distribution, and allows the intercept 'a0' and standard deviation 'std' to shift as state dependent parameters.}

\item{markovModel}{a \code{markov.model.object} from \code{select.markovModel}. If blank, the default selects a homogeneous Markov model without flickering.}
}
\value{
A built hydroState model object ready to be fitted with \code{fitModel}
}
\description{
\code{buildModel} builds a hydrostate model with either a default \code{stateModel} or the \code{stateModel} can be specified with options from \code{select.stateModel}. After the model is built, the hydroState model is ready to be fitted with \code{fitModel}
}
\details{
\code{buildModel}


hydroState operates in S4, object oriented programming, and requires three objects to build a hydroState model. Each object can be selected from \code{select.transform}, \code{select.stateModel}, and \code{select.markovModel}; however, if no object is defined a default model is built.

For an example of how to.. see vignette...
}
\keyword{build}
\keyword{hydroState}
