% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{checkModel}
\alias{checkModel}
\title{Check Model}
\usage{
checkModel(model, n.samples)
}
\arguments{
\item{model}{fitted \code{hydroState} model.}

\item{n.samples}{integer of samples to re-sample. Default is 100000.}
}
\value{
table of the probability of the inferred state equals the known state
}
\description{
\code{checkModel} validates the model's states at each time-step through re-sampleing the input data and re-runnning the Viterbi algorithm.
}
\details{
\code{checkModel}


Duplicate the input data 100 times. Get the transformed observations. Generate a synthetic series from the model. Get a time series of the transformed observations using the sample states. Find the Viterbi states for the re-sampled transformed data. Assess probability that the inferred state equals the 'known' state.
}
\examples{
## Check fitted model
checkModel(model = model.annual.fitted.221201)

}
\keyword{check}
\keyword{model}
\keyword{viterbi}
