% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{exportResiduals}
\alias{exportResiduals}
\title{Export residuals}
\usage{
exportResiduals(model.name = model)
}
\arguments{
\item{model.name}{is the name of the fitted hydroState model object.}
}
\value{
A matrix of residuals
}
\description{
export residuals
}
\details{
\code{exportResiduals}


\code{exportResiduals} exports the residuals of the fitted hydroState model. Ensure the figure pane is large enough to accommodate the size of the plot, else consider outputting as an image or pdf.
}
\keyword{export}
\keyword{residuals}
