% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{exportStates}
\alias{exportStates}
\title{Export states}
\usage{
exportStates(model.name = model)
}
\arguments{
\item{model.name}{is the name of the fitted hydroState model object.}
}
\value{
A matrix of results with observations and information for each state
}
\description{
exports results
}
\details{
\code{exportStates}


\code{exportStates} exports the results of the fitted hydroState model as a matrix. Columns include: year, Viterbi state number, mean and 90pct confidence intervals of Viterbi and Normal state values, conditional probabilities, and emission density of each state. This function does not plot states.
}
\keyword{Viterbi}
\keyword{export}
\keyword{results}
\keyword{states}
