% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{fitModel}
\alias{fitModel}
\title{Fit hydroState model}
\usage{
fitModel(model.name = model, pop.size.perParameter = 10, max.generations = 500)
}
\arguments{
\item{model.name}{name of the built hydroState model or name of the list containing all built models}

\item{pop.size.perParameter}{integer that should be greater than or equal to the number of parameters in the model. The default is '10' and is sufficient for all models.}

\item{max.generations}{integer that will stop the optimizer when set number of generations are reached. The default is '500'.}
}
\value{
A fitted hydroState model
}
\description{
\code{fitModel} fits hydrostate model(s) using global optimization by differential evolution \link[DEoptim]{DEoptim}.
}
\details{
\code{fitModel}


After a hydroState model object is built, the model is ready to be fitted. The only required input is the given name of the built hydroState model object. \code{fitModel} works for one built model (\code{buildModel}) or all (\code{buildModelAll}). If fitting all models be sure to install and load the \link[parallelly]{parallelly} library.
}
\keyword{fit}
\keyword{hydroState}
