% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{get.AIC}
\alias{get.AIC}
\title{get AIC}
\usage{
get.AIC(model, sample.penalty = FALSE)
}
\arguments{
\item{model}{fitted \code{hydroState} model object.}

\item{sample.penalty}{TRUE/FALSE. TRUE finds AICc, FALSE finds AIC}
}
\value{
AIC or AICc value(s)
}
\description{
\code{get.AIC} retrieves Akaike information criteria from a fitted model.
}
\details{
\code{get.AIC}


The AIC is the negative log-likelihood of the model plus a penealty for model parameters. This function can be performed on a single model or a selection of models to find the lowest AIC of the set. Set the `sample.penalty` to TRUE for finding the AICc.
}
\examples{
# Load fitted model
data(model.annual.fitted.221201)

## AIC of a single model
get.AIC(model.annual.fitted.221201)

## Lowest AIC of a model set
data(all.models.annual.fitted.407211)

get.AIC(all.models.annual.fitted.407211)


}
\keyword{AIC}
\keyword{AICc}
