% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{get.states}
\alias{get.states}
\title{get states}
\usage{
get.states(model.name = model)
}
\arguments{
\item{model.name}{is the name of the fitted \code{hydroState} model object.}
}
\value{
data frame of results to evaluate the rainfall-runoff states overtime
}
\description{
\code{get.states} retrieves results from the fitted \code{hydroState} model.
}
\details{
\code{get.states}


\code{get.states} These results include the time-step (i.e. Year, Month), Viterbi State Number to differentiate states, observations (i.e. stream flow), flow values of the Viterbi state including the 5\% and 95\% confidence intervals, flow values of the normal state including the 5% and 95% confidence intervals, conditional probabilities for each state, and emission density for each state. The Viterbi state flow values are the results, the values of the most likely state at each time-step. The Normal state flow values are the values from the normal state at each time-step. When the most likely state is the Normal state for a time-step, the Viterbi flow state value equals the Normal flow state value. These Normal state values are interpreted as the results for a one-state model, if the multi-state model did not exist. This estimated Normal state can be seen in plot (B) relative to the other states. The Conditional Probabilities of each state show the likelihood of remaining in the given state based on previous states. When the conditional probability is closer to 1, there is a higher probability that hydroState remains in the state for the next time-step. The Emission Density of each state is the result of multiplying the conditional probabilities by the transition probabilities.
}
\keyword{get}
\keyword{results}
\keyword{states}
