% These are internal functions not intended for the user
\name{hydroState-internal}
\alias{hydroState-internal}
\title{hydroState-internal}

\description{These are undocumented internal functions not intended for the user}

\alias{Qhat.boxcox}
\alias{Qhat.burbidge}
\alias{Qhat.log}
\alias{Qhat.none}
\alias{QhatModel.homo.gamma.linear}
\alias{QhatModel.homo.gamma.linear.AR1}
\alias{QhatModel.homo.gamma.linear.AR2}
\alias{QhatModel.homo.gamma.linear.AR3}
\alias{QhatModel.homo.normal.linear}
\alias{QhatModel.homo.normal.linear.AR1}
\alias{QhatModel.homo.normal.linear.AR2}
\alias{QhatModel.homo.normal.linear.AR3}
\alias{QhatModel.homo.skewedNormal.linear}
\alias{QhatModel.homo.skewedNormal.linear.AR1}
\alias{QhatModel.homo.skewedNormal.linear.AR2}
\alias{QhatModel.homo.skewedNormal.linear.AR3}
\alias{QhatModel.subAnnual.homo.gamma.linear}
\alias{QhatModel.subAnnual.homo.gamma.linear.AR1}
\alias{QhatModel.subAnnual.homo.gamma.linear.AR2}
\alias{QhatModel.subAnnual.homo.gamma.linear.AR3}
\alias{check.PseudoResiduals}
\alias{check.viterbi}
\alias{drought.resilience.index}
\alias{fit}
\alias{forecast}
\alias{getAIC}
\alias{getAIC.bestModel}
\alias{getAICc}
\alias{getBounds}
\alias{getBounds.transformed}
\alias{getBounds.transformed.asVector}
\alias{getDistributionPercentiles}
\alias{getEmissionDensity}
\alias{getInitialStateProbabilities}
\alias{getLogForwardProbabilities}
\alias{getLogLikelihood}
\alias{getNegLogLikelihood}
\alias{getNegLogLikelihood.fromTransformedVector}
\alias{getParameters}
\alias{getParameters.asVector}
\alias{getParameters.transformed}
\alias{getParameters.transformed.asVector}
\alias{getQ.backTransformed}
\alias{getQhat}
\alias{getStartEndIndex}
\alias{getStateFlicker}
\alias{getTransitionProbabilities}
\alias{hydroState.allModels}
\alias{hydroState.subAnnual.allModels}
\alias{markov.annualHomogeneous}
\alias{markov.annualHomogeneous.flickering}
\alias{parameters}
\alias{plot_graph}
\alias{setParameters}
\alias{setParameters.fromTransformed}
\alias{setParameters.fromTransformed.asVector}
\alias{setStateNames}
\alias{streamflow_annual_221201}
\alias{streamflow_annual_407211}
\alias{streamflow_gaugeInfo}
\alias{streamflow_monthly_221201}
\alias{streamflow_daily_221201}
\alias{model.annual.fitted.221201}
\alias{model.monthly.adjusted.fitted}
\alias{all.models.monthly.fitted.221201}
\alias{get.summary.table}
\alias{viterbi}
\alias{select.stateModel}
\alias{select.Markov}
\alias{select.transform}
