% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{plot.residuals}
\alias{plot.residuals}
\title{Plot residuals}
\usage{
\method{plot}{residuals}(model.name = model, do.pdf = FALSE, ID = NULL)
}
\arguments{
\item{model.name}{name of the fitted hydroState model object.}

\item{do.pdf}{option to export residual plots as a pdf. Default is FALSE.}

\item{ID}{character string of catchment identifier (i.e. gauge ID). Default is NULL. Only recommended when do.pdf = TRUE.}
}
\value{
Plots of residuals to evaluate model fit
}
\description{
The normal pseudo residuals are plotted for review to check for outliers and validate the fit of the model. It is recommended to ensure the model fit is valid before evaluating results (i.e. \code{plot.states}). Furthermore, to ensure the multi-state model performs better than the one-state model, it is recommended to visually compare \code{plot.resdiuals} of both models.
}
\details{
\code{plot.residuals}


\code{plot.residuals} produces five plots to review and validate the fitted hydroState model. A) Time-series of normal-pseudo residuals to ensure the residuals each year are within the confidence intervals. B) Auto-correlation function (ACF) of normal-pseudo residuals to ensure there is no serial correlation in residuals. Lag spikes should be below confidence interval at each lag (except 0). C) Histogram of uniform-pseudo residuals should show uniform distribution (equal frequency for each residual value) D) Histogram of normal-pseudo residuals should show normal distribution centered on zero and with no skew E) Quantile-Quantile (Q-Q) plot where normal-pseudo residuals vs. theoretical quantities should align on the diagonal line. The last plot contains the Akaike information criterion (AIC) and Shapiro-Wilk p-value. The AIC is an estimator to determine the most parsimonious, best performing model given the number of parameters. When comparing models, the lowest AIC is the best performing model. Shapiro-Wilks test for normality in the residuals and a p-value greater than 0.05 (chosen alpha level) indicates the residuals are normally distributed; the null hypothesis that the residuals are normally distributed is not rejected.
}
\keyword{plot}
\keyword{residuals}
