% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{plot.states}
\alias{plot.states}
\title{plot States}
\usage{
\method{plot}{states}(
  model.name = model,
  ind.variable = TRUE,
  dep.variable = TRUE,
  dep.variable.transformed = TRUE,
  cond.state.prob = TRUE,
  do.pdf = FALSE,
  ID = NULL
)
}
\arguments{
\item{model.name}{is the name of the fitted hydroState model object.}

\item{ind.variable}{option to plot independent variable overtime. Default is TRUE.}

\item{dep.variable}{option to plot dependent variable and states overtime. Default is TRUE.}

\item{dep.variable.transformed}{option to plot transformed dependent variable and states overtime. Default is TRUE.}

\item{cond.state.prob}{option to plot the conditional state probabilities overtime for each state. Default is TRUE.}

\item{do.pdf}{option to export plots as a pdf. Default is FALSE.}

\item{ID}{character string of catchment identifier (i.e. gauge ID). Default is NULL. Only recommended when do.pdf = TRUE.}
}
\value{
plots to evaluate rainfall-runoff states overtime along with observations and the conditional probabilities of each state. The data frame includes several items. These include the time-step (i.e. Year, Month), Viterbi State Number to differentiate states, observations (i.e. stream flow), flow values of the Viterbi state including the 5\% and 95\% confidence intervals, flow values of the normal state including the 5% and 95% confidence intervals, conditional probabilities for each state, and emission density for each state. The Viterbi state flow values are the results, the values of the most likely state at each time-step. The Normal state flow values are the values from the normal state at each time-step. When the most likely state is the Normal state for a time-step, the Viterbi flow state value equals the Normal flow state value. These Normal state values are interpreted as the results for a one-state model, if the multi-state model did not exist. This estimated Normal state can be seen in plot (B) relative to the other states. The Conditional Probabilities of each state show the likelihood of remaining in the given state based on previous states. When the conditional probability is closer to 1, there is a higher probability that hydroState remains in the state for the next time-step. The Emission Density of each state is the result of multiplying the conditional probabilities by the transition probabilities.
}
\description{
\code{plot.states} produces several plots to visualize results of the states overtime. \code{setInitialYear} is required before \code{plot.states}.
}
\details{
\code{plot.states}


\code{plot.states} produces plots of the results from the fitted hydroState model. The default produces four plots of the A) independent variable, B) dependent variable and states, C) transformed dependent variable and states, and D) conditional state probabilities for each state. These are plotted on the same page, and there is an option to export plots as a pdf to the current working directory. There are also options to only plot one of the four plots.
}
\keyword{plot}
\keyword{results}
\keyword{states}
