% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{plotResiduals}
\alias{plotResiduals}
\title{Plot residuals}
\usage{
plotResiduals(model.name = model)
}
\arguments{
\item{model.name}{is the name of the fitted hydroState model object.}
}
\value{
A plot of residuals
}
\description{
Plots residuals
}
\details{
\code{plotResiduals}


\code{plotResiduals} plots the residuals of the fitted hydroState model. Five plots are provided to visualize the residuals: A) time-series of normal-pseudo residuals, B) auto-correlation of normal-pseudo residuals, C) histogram of uniform-pseudo residuals, D) histogram of normal-pseudo residuals, and E) Q-Q plot of normal-pseudo residuals
}
\keyword{plot}
\keyword{residuals}
