% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{plotStates}
\alias{plotStates}
\title{Plot states}
\usage{
plotStates(model.name = model)
}
\arguments{
\item{model.name}{is the name of the fitted hydroState model object.}
}
\value{
A plot of results with observations and states
}
\description{
Plots results overtime
}
\details{
\code{plotStates}


\code{plotStates} plots the results of the fitted hydroState model. Five plots are provided to visualize results at each time-step: A) observed precipitation, B) observed runoff with states, C) transformed runoff with states, D) cumulative rainfall residual, and E) the state probabilities (emission density). Ensure the figure pane is large enough to accommodate the size of the plot, else consider outputting as an image or pdf.
}
\keyword{names}
\keyword{plot}
\keyword{results}
\keyword{states}
