% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{select.markovModel}
\alias{select.markovModel}
\title{\code{select.markovModel} selects a hidden Markov model}
\usage{
select.markovModel(
  func = "annualHomogeneous",
  transition.graph = matrix(TRUE, 2, 2)
)
}
\arguments{
\item{func}{is the Markov model defining the transition between states. The default is 'annualHomogeneous'. Other options include 'annualHomogeneous.flickering'.}

\item{transition.graph}{is a matrix given the number of states. The default is a 2-state matrix (2 by 2): matrix(TRUE,2,2)}
}
\value{
A Markov object with a selected Markov model
}
\description{
select.markovModel selects a hidden Markov model.
}
\details{
There are several hidden Markov models to choose from with different forms. The default markov model is 'annualHomogeneous' where the transition between states only depends on a sequence of observed prior states.


For an example of how to.. see vignette...
}
\keyword{HMM}
\keyword{Markov}
\keyword{homogeneous}
