% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{select.transform}
\alias{select.transform}
\title{Transforms Observations}
\usage{
select.transform(
  func = "boxcox",
  input.data = data.frame(year = c(), flow = c(), precip = c())
)
}
\arguments{
\item{func}{is the method of transformation. The default is 'boxcox'. Other options: 'log', 'burbidge', 'none'}

\item{input.data}{dataframe of annual, seasonal, or monthly runoff and precipitation observations. Gaps with missing data in either streamflow or precipitation are permitted, and the handling of them is further discussed in \code{select.Markov}. Monthly data is required when using \code{seasonal.parameters} that assumes selected model parameters are better defined with a sinusoidal function.}
}
\value{
A Qhat object with transformed observations ready for \code{buildModel}
}
\description{
Transforms observations to remove heteroscedasticity.
}
\details{
\code{select.transform}


Often there is skew within hydrologic data. When defining relationships between observations, this skew results in an unequal variance in the residuals, heteroscedasticity. Transforming observations is often required with observations of streamflow, precipitation, and concentration. Qhat provides several options to transform observations. Since the degree of transformation is not typically known, 'boxcox' is the default. Other options include: 'log', 'burbidge', and of course, 'none' when no transformation is performed.
}
\examples{
# Load data
data(streamflow_annual_221201)

# Transform observations
data.transform = select.transform(func = 'boxcox',input.data = streamflow_annual_221201)
#or
data.transform = select.transform(func = 'log',input.data = streamflow_annual_221201)
#or
data.transform = select.transform(func = 'burbidge',input.data = streamflow_annual_221201)
#or
data.transform = select.transform(func = 'none',input.data = streamflow_annual_221201)

}
\keyword{Qhat}
\keyword{transform}
