% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{setInitialYear}
\alias{setInitialYear}
\title{Sets state names given initial year}
\usage{
setInitialYear(model.name = model, initial.year = streamflow_annual$hy_year[1])
}
\arguments{
\item{model.name}{is the name of the fitted hydroState model object.}

\item{initial.year}{is a year (YYYY). Default is first year in input.data.}
}
\value{
A fitted hydroState model object with state names for each time-step
}
\description{
sets the state names for each time-step relative to the initial year given
}
\details{
\code{setInitialYear}


hydroState assigns names to the computed states. This requires choosing an initial year where the state value from that year will be named 'Normal'. Other state values will be given names relative to the state value in the initial year. The choice of the initial year does not affect results. It is really just a means to more easily interpret the difference in state values relative to each other. It is best to choose a year based on the question being asked. For example, in testing the impact of drought, a year before the beginning of the drought was selected as an initial year when conditions were condsidered 'Normal' (Peterson TJ, Saft M, Peel MC & John A (2021), Watersheds may not recover from drought, Science, DOI: \doi{10.1126/science.abd5085})
}
\keyword{names}
\keyword{state}
