% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{showModelAll}
\alias{showModelAll}
\title{Show all models}
\usage{
showModelAll(all.models)
}
\arguments{
\item{all.models}{hydroState.allModels object with a list of models from \code{buildModelAll}}
}
\value{
A data.frame with a summary of all models
}
\description{
\code{showModelAll} outputs a summary table of all built hydroState models and allows users to edit the reference models for calibration.
}
\details{
\code{showModelAll}


For every model in code{buildModelAll}, there is a reference model for calibration. The reference model is slightly simpler model with one less parameter. During calibration with \code{fitModel}, the model performance must exceed the the performance of the reference model else the model is rejected. For instance 'model.1State.normal.log.AR0' contains 3-parameters and is the reference model for 'model.1State.normal.log.AR1' which contains 4-parameters. The objective function of 'model.1State.normal.log.AR1' must calibrate the model with a lower negative log-likelihood than 'model.1State.normal.log.AR0'. These reference models are pre-defined in \code{buildModelAll}, but this function allows the user to edit the reference models in the data.frame if needed.
}
\keyword{all}
\keyword{hydroState}
\keyword{models}
\keyword{summary}
