% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{buildModelAll}
\alias{buildModelAll}
\title{Builds all hydroState models # work in progress..}
\usage{
buildModelAll(
  ID = "",
  input.data = data.frame(year = c(), flow = c(), precip = c())
)
}
\arguments{
\item{ID}{character vector of a stream gauge identifier}

\item{input.data}{is a dataframe of annual, monthly, or daily, observations of runoff, precipitation, or concentration. For running seasonal models, monthly values are required.}
}
\value{
A list of built hydroState models with every combination of QhatModel, ready to be fitted
}
\description{
\code{buildModelAll} builds all hydroState models
}
\details{
\code{buildModelAll}


All hydroState models are build with different QhatModel objects

For an example of how to.. see vignette...
}
\keyword{all}
\keyword{build}
\keyword{hydroState}
