% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{get.AIC}
\alias{get.AIC}
\title{get AIC}
\usage{
get.AIC(model.name)
}
\arguments{
\item{model.name}{is the name of the fitted \code{hydroState} model object.}
}
\value{
AIC value
}
\description{
\code{get.AIC} retrieves Akaike information criteria from a fitted model.
}
\details{
\code{get.AIC}


The AIC is the negative log-likelihood of the model plus a penealty for model parameters. This function can be performed on a single model or a selection of models to find the lowest AIC of the set.
}
\examples{
# Load fitted model
data(model.annual.fitted.221201)

## AIC of a single model
get.AIC(model.annual.fitted.221201)

## Lowest AIC of a model set
# get.AIC()

}
\keyword{AIC}
