% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{get.residuals}
\alias{get.residuals}
\title{Get residuals}
\usage{
get.residuals(model.name = model)
}
\arguments{
\item{model.name}{name of the fitted hydroState model object.}
}
\value{
Data frame of residuals for each time-step
}
\description{
The normal pseudo residuals are retrieved from the fitted model.
}
\details{
\code{get.residuals}


\code{get.residuals} retrieves residuals from the fitted model and exports them as a data frame.
}
\examples{
# Load fitted model
data(model.annual.fitted)

## Get residuals in a dataframe
get.residuals(model.name = model.annual.fitted)


}
\keyword{residuals}
