% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{plot.residuals}
\alias{plot.residuals}
\title{Plot residuals}
\usage{
\method{plot}{residuals}(model, do.pdf = FALSE, ID = NULL)
}
\arguments{
\item{model}{fitted hydroState model object.}

\item{do.pdf}{option to export residual plots as a pdf. Default is FALSE.}

\item{ID}{character string of catchment identifier (i.e. gauge ID). Default is NULL. Only recommended when do.pdf = TRUE.}
}
\value{
Plots of residuals to evaluate model fit
}
\description{
The normal pseudo residuals are plotted for review to check for outliers and validate the fit of the model. It is recommended to ensure the model fit is valid before evaluating results (i.e. \code{plot.states}). Furthermore, to ensure the multi-state model performs better than the one-state model, it is recommended to visually compare \code{plot.resdiuals} of both models.
}
\details{
\code{plot.residuals}


\code{plot.residuals} produces five plots to review and validate the fitted hydroState model.
\itemize{
 \item{A)}{ Time-series of normal-pseudo residuals to ensure the residuals each year are within the confidence intervals.}
 \item{B)}{ Auto-correlation function (ACF) of normal-pseudo residuals to ensure there is no serial correlation in residuals. Lag spikes should be below confidence interval at each lag (except 0).}
 \item{C)}{ Histogram of uniform-pseudo residuals should show uniform distribution (equal frequency for each residual value)}
 \item{D)}{ Histogram of normal-pseudo residuals should show normal distribution centered on zero and with no skew}
 \item{E)}{ Quantile-Quantile (Q-Q) plot where normal-pseudo residuals vs. theoretical quantities should align on the diagonal line. The last plot contains the Akaike information criterion (AIC) and Shapiro-Wilk p-value. The AIC is an estimator to determine the most parsimonious, best performing model given the number of parameters. When comparing models, the lowest AIC is the best performing model. Shapiro-Wilks test for normality in the residuals and a p-value greater than 0.05 (chosen alpha level) indicates the residuals are normally distributed; the null hypothesis that the residuals are normally distributed is not rejected.}
 }
 It is recommended to export the residual plot as a PDF due to it's size. If the R plot windor is too small, two common errors can occur:
 \itemize{
 \item{"Error in plot.new() : figure margins too large":} reset plot window with "dev.off()", enlarge plot area and re-run \code{plot.residuals}.
 \item{"Error in par(op) : invalid value specified for graphical parameter "pin"} if the R plot window is not reset with "dev.off", an additional \code{plot.residuals} attempt will result in this error.
 }
}
\examples{
# Load fitted model
data(model.annual.fitted.221201)

## Plot residuals
plot.residuals(model = model.annual.fitted.221201)

}
\keyword{plot}
\keyword{residuals}
