% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{plot.states}
\alias{plot.states}
\title{plot States}
\usage{
\method{plot}{states}(
  model.name,
  ind.variable = TRUE,
  dep.variable = TRUE,
  dep.variable.transformed = TRUE,
  cond.state.prob = TRUE,
  do.pdf = FALSE,
  ID = NULL
)
}
\arguments{
\item{model.name}{is the name of the fitted hydroState model object.}

\item{ind.variable}{option to plot independent variable overtime. Default is TRUE.}

\item{dep.variable}{option to plot dependent variable and states overtime. Default is TRUE.}

\item{dep.variable.transformed}{option to plot transformed dependent variable and states overtime. Default is TRUE.}

\item{cond.state.prob}{option to plot the conditional state probabilities overtime for each state. Default is TRUE.}

\item{do.pdf}{option to export plots as a pdf. Default is FALSE.}

\item{ID}{character string of catchment identifier (i.e. gauge ID). Default is NULL. Only recommended when do.pdf = TRUE.}
}
\value{
plots to evaluate rainfall-runoff states overtime along with observations and the conditional probabilities of each state.
}
\description{
\code{plot.states} produces several plots to visualize results of the states overtime. \code{setInitialYear} is required before \code{plot.states}.
}
\details{
\code{plot.states}


\code{plot.states} produces four figures of the results from the fitted hydroState model. The default produces all four:
\itemize{
 \item{independent variable}: precipitation
 \item{dependent variable and states}: streamflow observations, most likely state, and relative normal state estimate
 \item{transformed dependent variable and states}: transformed streamflow observations and most likely state
 \item{conditional state probabilities for each state}: probability of hydroState model remaining in given state
 }
 These are plotted on the same page, and there is an option to export plots as a pdf to the current working directory. There are also options to only plot one of the four plots.
}
\examples{
# Load fitted model
data(model.annual.fitted.221201)

## Set initial year to set state names
model.annual.fitted.221201 =
                  setInitialYear(model.name = model.annual.fitted.221201,
                  initial.year = 1990)

## Plot all figures
plot.states(model.name = model.annual.fitted.221201)

## Plot only dependent variable transformed with markov states
plot.states(model.name = model.annual.fitted.221201,
             ind.variable = FALSE,
             dep.variable = FALSE,
             dep.variable.transformed = TRUE,
             cond.state.prob = FALSE)

}
\keyword{plot}
\keyword{results}
\keyword{states}
