% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{select.transform}
\alias{select.transform}
\title{Transforms Observations}
\usage{
\method{select}{transform}(
  func = "boxcox",
  input.data = data.frame(year = c(), flow = c(), precip = c())
)
}
\arguments{
\item{func}{is the method of transformation. The default is 'boxcox'. Other options: 'log', 'burbidge', 'none'}

\item{input.data}{dataframe of annual or monthly runoff and precipitation observations.}
}
\value{
A Qhat object with transformed observations ready for \code{buildModel}
}
\description{
Transforms observations to remove heteroscedasticity.
}
\details{
\code{select.transform}


Often there is skew within hydrologic data. When defining relationships between observations, this skew results in an unequal variance in the residuals, heteroscedasticity. Transforming observations is often required with observations of streamflow, precipitation, and concentration. Qhat provides several options to transform observations. Since the degree of transformation is not typically known, 'boxcox' is the default. Other options include: 'log', 'burbidge', and of course, 'none' when no transformation is performed.

For an example of how to transform observations.. see vignette...
}
\keyword{Qhat}
\keyword{transform}
